refactor(reef): extract a presentational MCP clients list
The settings page's MCP client table is the UI onboarding needs for its
"connect your agent over MCP" step, so pull it out as McpClientsList in
components/mcp-clients-list. It takes an optional maxHeight and sets its own
overflow inline, so bounding it scrolls the rows under the sticky header
without a caller having to outrank the container's own `overflow: hidden`.

The heading, docs link, and banner move to the settings route rather than
travelling with the list: that copy is page specific, down to the words
"this page". Deleting the wrapper leaves the folder named after the one
component in it.

Covers the list with its first tests: row labels, picking and clearing a
workspace, the pending lock, each status message, and the bounded scroll.
