+++ apps/reef/app/components/mcp-clients-list/index.ts
@@ -1,0 +1,6 @@
+export {
+  McpClientsList,
+  type McpClientListItem,
+  type McpClientsListProps,
+} from './mcp-clients-list'
+export { type DesktopMcpClientsState, useDesktopMcpClients } from './use-desktop-mcp-clients'
