+++ apps/reef/app/components/mcp-clients-list/mcp-clients-list.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta, StoryObj } from '@storybook/react-vite'
 import { fn } from 'storybook/test'
 
-import { McpClientsSettings } from './mcp-clients-settings'
+import { McpClientsList } from './mcp-clients-list'
 
 const meta = {
   args: {
@@ -12,28 +12,34 @@ const meta = {
     onWorkspaceChange: fn(),
     workspaces: [{ name: 'default' }, { name: 'analytics' }],
   },
-  component: McpClientsSettings,
+  component: McpClientsList,
   parameters: { layout: 'padded' },
   render: (args) => (
     <div style={{ maxWidth: 960 }}>
-      <McpClientsSettings {...args} />
+      <McpClientsList {...args} />
     </div>
   ),
   tags: ['autodocs'],
-  title: 'Components/McpClientsSettings',
-} satisfies Meta<typeof McpClientsSettings>
+  title: 'Components/McpClientsList',
+} satisfies Meta<typeof McpClientsList>
 
 export default meta
 type Story = StoryObj<typeof meta>
 
 export const Default: Story = {}
 
-export const Loading: Story = {
-  args: { clients: [], loading: true },
-}
-
-export const Empty: Story = {
-  args: { clients: [] },
+export const Bounded: Story = {
+  args: {
+    clients: ['Claude Code', 'Claude Desktop', 'Codex', 'Cursor', 'VS Code', 'Zed'].map(
+      (name, index) => ({
+        configuredWorkspace: index === 0 ? 'default' : undefined,
+        id: name.toLowerCase().replaceAll(' ', '-'),
+        name,
+      }),
+    ),
+    maxHeight: 200,
+    pendingClientIds: ['codex'],
+  },
 }
 
 export const Error: Story = {
