--- apps/reef/app/components/mcp-clients-settings/index.ts
@@ -1,6 +1,0 @@
-export {
-  McpClientsSettings,
-  type McpClientSettingsItem,
-  type McpClientsSettingsProps,
-} from './mcp-clients-settings'
-export { useDesktopMcpClients } from './use-desktop-mcp-clients'
